fix(Table): 修复 verticalLine 打开时，多级表头的部分 th 右边框显示问题
